rename warningConfig to warningsConfig and remove warning prefix for messages in warningConfig
