+++ packages/admin/cms-admin/src/warnings/warningsConfig.ts
@@ -1,0 +1,14 @@
+import { type ReactNode } from "react";
+
+import { useCometConfig } from "../config/CometConfigContext";
+import { warningMessages as cometWarningMessages } from "./warningMessages";
+
+export interface WarningsConfig {
+    messages: Record<string, ReactNode>;
+}
+
+export function useWarningConfig(): WarningsConfig {
+    const cometConfig = useCometConfig();
+
+    return { ...cometConfig.warnings, messages: { ...cometConfig.warnings?.messages, ...cometWarningMessages } };
+}
